move queries to query folder
