+++ src/app/generate-queries/gen-queries.spec.ts
@@ -13,10 +13,10 @@ import {
   TestBed,
 } from '@angular/core/testing'; // tslint:disable-line
 
-import { localLogout } from './authentication/auth';
-import { GraphApiVersion, GraphApiVersions, IGraphRequestHeader, substituteTokens } from './base';
+import { GraphApiVersion, GraphApiVersions, IGraphRequestHeader, substituteTokens } from '../base';
+import { localLogout } from './../authentication/auth';
+import { GraphService } from './../graph-service';
 import { SampleQueries } from './gen-queries';
-import { GraphService } from './graph-service';
 
 function getGraphVersionFromUrl(url: string): GraphApiVersion {
   for (const version of GraphApiVersions) {
@@ -111,18 +111,18 @@ describe('Sample query validation', () => {
 
       graphService.performAnonymousQuery(query.method, 'https://graph.microsoft.com' + query.requestUrl, headers)
         .then((res) => {
-        if (res.headers.get('Content-Type').indexOf('application/json') !== -1) {
-          const response = res.json();
-          if (response && response.value && response.value.constructor === Array) {
-            if (response.value.length === 0 && skipResponseLengthCheck()) {
-              done.fail(`${query.humanName}: All sample GETs on collections must have values`);
+          if (res.headers.get('Content-Type').indexOf('application/json') !== -1) {
+            const response = res.json();
+            if (response && response.value && response.value.constructor === Array) {
+              if (response.value.length === 0 && skipResponseLengthCheck()) {
+                done.fail(`${query.humanName}: All sample GETs on collections must have values`);
+              }
             }
           }
-        }
-        done();
-      }).catch((e: Response) => {
-        done.fail(`${query.humanName}: Can't execute sample GET request, ${e.status}, ${JSON.stringify(e.json())}`);
-      });
+          done();
+        }).catch((e: Response) => {
+          done.fail(`${query.humanName}: Can't execute sample GET request, ${e.status}, ${JSON.stringify(e.json())}`);
+        });
     });
   }
 });
