+++ tslint.json
@@ -59,7 +59,7 @@
   "linterOptions": {
     "exclude": [
       "src/app/loc_strings.ts",
-      "src/app/gen-queries.ts"
+      "src/app/generate-queries/gen-queries.ts"
     ]
   }
 }
